Model Deployment/FastAPI/requirements.txt
tensorflow-cpu==2.5.0
fastapi
uvicorn
gunicorn==20.1.0
python-multipart
numpy==1.19.5
Pillow==9.0.1